Split competition estimator into separate files
